wtxmgr: move legacy tx wrappers
Move the touched legacy transaction helper entrypoints into
wtxmgr/deprecated.go without changing their behavior or docs.

This keeps the compatibility-facing wrappers separate from the
canonical implementation paths.
